requirements.txt
uvicorn==0.32.0
fastapi==0.115.3
asyncpg==0.30.0
passlib==1.7.4
python-multipart==0.0.17
pyjwt==2.9.0
redis==5.2.0
loguru==0.7.2
